Adding Bugzilla.py file which contains work in progress class which can interface with the Bugzilla 5.0 API for getting bug status's and updating bug data.
The goal of this is to eventually create a precommit hook which can
update teh status of a bug automatically from a commit message. To in
crease developer productivity. 

There once was such an extension for mercurial but unfortunately it
seems to be broken when using the Bugzilla 5.0 or later.

README.md has been updated with usage info if you would like to develop.